Define concrete as_dict() in base Config class
