+++ src/uwtools/config/formats/ini.py
@@ -80,12 +80,6 @@ class INIConfig(Config):
 
     # Public methods
 
-    def as_dict(self) -> dict:
-        """
-        Returns a pure dict version of the config.
-        """
-        return self.data
-
     def dump(self, path: Path | None = None) -> None:
         """
         Dump the config in INI format.
