+++ src/uwtools/config/formats/sh.py
@@ -80,12 +80,6 @@ class SHConfig(Config):
 
     # Public methods
 
-    def as_dict(self) -> dict:
-        """
-        Returns a pure dict version of the config.
-        """
-        return self.data
-
     def dump(self, path: Path | None) -> None:
         """
         Dump the config as key=value lines.
