+++ src/uwtools/config/formats/yaml.py
@@ -139,12 +139,6 @@ class YAMLConfig(Config):
 
     # Public methods
 
-    def as_dict(self) -> dict:
-        """
-        Returns a pure dict version of the config.
-        """
-        return self.data
-
     def dump(self, path: Path | None = None) -> None:
         """
         Dump the config in YAML format.
